WIP... Committing before switching approach
Currently it "works" in communicating with the first agent, but it returns an error. It doesnt look like the right input data is received (apparently we read it with pandas and send it instead of attaching it. Not sure if that is the way to go? We might have to change so it has the input file to run on with code interpreter.) Secondly we are not using code interpreter. So we have to change tools. And we may therefore have to change to using assistants? I hope not. Lets see.