BREAKTHORUGH. It is communicating with an agent
We got through. Although it does output an error (it tries to use a custom tool that does not exist or is not yet implemented for the agents uuse). However, we cannot and should not manually implement individual tools. We need to make something more versatile. Code writing to run? Maybe.